INTERNAL: Introduced class ColumnsInfo
the main motivation for the refactoring was the need to remove variable "columnName" from DA.Service code. The variable sometimes contained a single column name, sometimes several column names separated by ";" and it was not clear that the ";" could be there. The second reason is the need to let the AbstractSqlGenerator know which column names are associated with virtual columns (will be used in later commits).