INTERNAL: Certain columns no longer missing in the configuration dialog.
If no configuration was given the configuration dialog was filled in with default column set created from all panel properties.
When default configuration was given explicitly in screen xml and it did not contain some properties, these would never appear in the dialog, because the implicit default configuration was not taken into account.

The fix takes the explicit (stored by user) default configuration and merges it into the implicit (created from panel properties) default configuration. Where no user stored column configuration is not provided, the implicit default one is used.