INTERNAL: DestroyForm action result is processed before any other actions.
1) Fixes https://trello.com/c/fj4WDGwT - by altering the order of processed results the error is correctly associated with the screen which raised it, because the original dialog screen has been already closed.
2) Implements https://trello.com/c/COo60Ad8 - Will close the dialog first, then it starts opening some screen.

This would probably be better if fixed on server, but it is IMO just a matter of convention. If there is any DestroyForm in the results, that always should mean "Close the topmost form" - the one which was topmost when the sourcing action has been triggered. I see no point having a chance to issue DestroyForm after any other result type. So it shall not matter where this gets implemented.