INTERNAL: Fix infinite recursion when search result item is clicked.
The menu item is passed to immer produce function to immutably change it, but the item has been deserialized from xml with parent option turned on, causing it to be part of a cyclical structure.
Immutable structures in js-world sense cannot have cycles so produce function does not like them.