FIX: Filter UI conflict with add/copy row (#1235)
Creating a new row or duplicating an existing row while the filters were active led to filters to be closed and grid to be reloaded which caused active row to be switched to the first row because the dirty row was not returned by the GetRows call. This caused an error on the server.