FIX: react-virtualized (#1577)
* Revert "DEV: Client clean up (#1544)" - this reverts commit c530844a

The package react-virtualized 9.22.3 contains a bug. The bug was fixed during origam build using the patch-package features. We recently upgraded the react-virtualized to 9.22.5 which was a bug fix release. The patch-package fix was removed because it was assumed that the bug was fixed. But it turns out it was not. That is why the "DEV: Client clean up" commit was reverted.